Optimizing with numba also nonparametric occurrence rates. Here is an example with a source from SAM (speedup 9x in "computing pnes"):
```
# before
| calc_53715, maxmem=4.0 GB | time_sec | memory_mb | counts |
|---------------------------+----------+-----------+--------|
| total classical           | 394.9    | 90.4      | 23     |
| computing pnes            | 279.3    | 0.0       | 626    |
| ClassicalCalculator.run   | 70.7     | 206.4     | 1      |
| get_poes                  | 59.6     | 0.0       | 626    |
| computing mean_std        | 18.9     | 0.0       | 626    |
| make_contexts             | 14.2     | 24.4      | 23     |
# after
| calc_53738, maxmem=4.1 GB | time_sec | memory_mb | counts |
|---------------------------+----------+-----------+--------|
| total classical           | 151.7    | 90.7      | 23     |
| get_poes                  | 62.3     | 0.0       | 626    |
| ClassicalCalculator.run   | 39.7     | 198.9     | 1      |
| computing pnes            | 31.5     | 0.0       | 626    |
| computing mean_std        | 19.6     | 0.0       | 626    |
| make_contexts             | 14.6     | 24.2      | 23     |
```